Commit iOS refactored code
