+++ src/edu/umass/cs/gigapaxos/PaxosConfig.java
@@ -30,10 +30,6 @@ import java.util.logging.ConsoleHandler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import edu.umass.cs.gigapaxos.interfaces.ExecutedCallback;
-import edu.umass.cs.gigapaxos.paxospackets.AcceptPacket;
-import edu.umass.cs.gigapaxos.paxospackets.RequestPacket;
-
 import org.junit.Assert;
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -1160,31 +1156,4 @@ public class PaxosConfig {
 						: null;
 	}
 
-    /**
-     *
-     */
-     public static class RequestAndCallback {
-        /**
-         *
-         */
-         protected RequestPacket requestPacket;
-        /**
-         *
-         */
-         final ExecutedCallback callback;
-
-        RequestAndCallback(RequestPacket request, ExecutedCallback callback) {
-            this.requestPacket = request;
-            this.callback = callback;
-        }
-        /**
-         * @return RequestPacket
-         */
-        public RequestPacket getRequestPacket() {
-            return this.requestPacket;
-        }
-        protected AcceptPacket setAcceptPacket(AcceptPacket accept) {
-            return (AcceptPacket)(this.requestPacket= accept);
-        }
-    }
 }
