+++ src/edu/umass/cs/gigapaxos/PaxosManager.java
@@ -158,18 +158,46 @@ public class PaxosManager<NodeIDType> {
 	private static final boolean USE_GC_MAP = Config
 			.getGlobalBoolean(PC.USE_GC_MAP);
 
-	private class Outstanding {
+    /**
+     *
+     */
+     public static class RequestAndCallback {
+        /**
+         *
+         */
+         protected RequestPacket requestPacket;
+        /**
+         *
+         */
+         final ExecutedCallback callback;
+
+        RequestAndCallback(RequestPacket request, ExecutedCallback callback) {
+            this.requestPacket = request;
+            this.callback = callback;
+        }
+        /**
+         * @return RequestPacket
+         */
+        public RequestPacket getRequestPacket() {
+            return this.requestPacket;
+        }
+        protected AcceptPacket setAcceptPacket(AcceptPacket accept) {
+            return (AcceptPacket)(this.requestPacket= accept);
+        }
+    }
+
+    private class Outstanding {
 		int totalRequestSize = 0;
 		long lastIncremented = System.currentTimeMillis();
-		ConcurrentHashMap<Long, PaxosConfig.RequestAndCallback> requests = USE_GC_MAP ? new GCConcurrentHashMap<Long, PaxosConfig.RequestAndCallback>(
+		ConcurrentHashMap<Long, RequestAndCallback> requests = USE_GC_MAP ? new GCConcurrentHashMap<Long, RequestAndCallback>(
 				new GCConcurrentHashMapCallback() {
 					@Override
 					public void callbackGC(Object key, Object value) {
-						PaxosManager.this.callbackRequestTimeout(((PaxosConfig.RequestAndCallback) value).requestPacket);
+						PaxosManager.this.callbackRequestTimeout(((RequestAndCallback) value).requestPacket);
 					}
 				}, REQUEST_TIMEOUT)
-				: new ConcurrentHashMap<Long, PaxosConfig.RequestAndCallback>();
-		HashMap<RequestPacket, PaxosConfig.RequestAndCallback> conflictIDRequests = new HashMap<RequestPacket, PaxosConfig.RequestAndCallback>();
+				: new ConcurrentHashMap<Long, RequestAndCallback>();
+		HashMap<RequestPacket, RequestAndCallback> conflictIDRequests = new HashMap<RequestPacket, RequestAndCallback>();
 		ConcurrentHashMap<RequestPacket, RequestResponseAndCallback> responses = USE_GC_MAP ? new GCConcurrentHashMap<RequestPacket, RequestResponseAndCallback>(
 				new GCConcurrentHashMapCallback() {
 					@Override
@@ -178,10 +206,10 @@ public class PaxosManager<NodeIDType> {
 				}, REQUEST_TIMEOUT)
 				: new ConcurrentHashMap<RequestPacket, RequestResponseAndCallback>();
 
-		private void enqueue(PaxosConfig.RequestAndCallback rc) {
+		private void enqueue(RequestAndCallback rc) {
 			assert (rc.requestPacket.getType() != PaxosPacketType.ACCEPT || rc.requestPacket
 					.hasRequestValue());
-			PaxosConfig.RequestAndCallback prev = null;
+			RequestAndCallback prev = null;
 			synchronized (this.requests) {
 				if ((prev = this.requests.putIfAbsent(rc.requestPacket.requestID, rc)) == null)
 					totalRequestSize += rc.requestPacket.lengthEstimate();
@@ -198,8 +226,8 @@ public class PaxosManager<NodeIDType> {
 		}
 
 		// called by executed callback
-		private PaxosConfig.RequestAndCallback dequeue(RequestPacket request) {
-			PaxosConfig.RequestAndCallback queued = this.requests.get(request.requestID);
+		private RequestAndCallback dequeue(RequestPacket request) {
+			RequestAndCallback queued = this.requests.get(request.requestID);
 			if (queued != null && queued.requestPacket.equals(request))
 				return this.requests.remove(request.requestID);
 			else
@@ -221,7 +249,7 @@ public class PaxosManager<NodeIDType> {
 
 	private void GC() {
 		if (this.outstanding.requests instanceof GCConcurrentHashMap)
-			((GCConcurrentHashMap<Long, PaxosConfig.RequestAndCallback>) this.outstanding.requests)
+			((GCConcurrentHashMap<Long, RequestAndCallback>) this.outstanding.requests)
 					.tryGC(REQUEST_TIMEOUT);
 		else if (System.currentTimeMillis() - this.outstanding.lastIncremented > PaxosManager.FADE_OUTSTANDING_TIMEOUT) {
 			if (this.outstanding.requests.size() > MAX_OUTSTANDING_REQUESTS)
@@ -231,7 +259,7 @@ public class PaxosManager<NodeIDType> {
 		}
 	}
 
-	static class RequestResponseAndCallback extends PaxosConfig.RequestAndCallback {
+	static class RequestResponseAndCallback extends RequestAndCallback {
 		/* this is actually the request whose getResponse contains the response. */
 		final Request clientRequest;
 
@@ -282,7 +310,7 @@ public class PaxosManager<NodeIDType> {
 	// called by PaxosInstanceStateMachine as execute callback
 	protected boolean executed(RequestPacket requestPacket, Request request,
 			boolean sendResponse) {
-		PaxosConfig.RequestAndCallback rc = this.outstanding.dequeue(requestPacket);
+		RequestAndCallback rc = this.outstanding.dequeue(requestPacket);
 		if (rc != null)
 			this.outstanding.totalRequestSize -= rc.requestPacket.lengthEstimate();
 		RequestInstrumenter.remove(requestPacket.requestID);
@@ -305,7 +333,7 @@ public class PaxosManager<NodeIDType> {
 		RequestResponseAndCallback rrc = null;
 		if (ENABLE_RESPONSE_CACHING
 				&& (rrc = this.outstanding.responses.get((requestPacket))) != null) {
-			PaxosConfig.RequestAndCallback rc = this.outstanding.dequeue(requestPacket);
+			RequestAndCallback rc = this.outstanding.dequeue(requestPacket);
 			if (rc!=null && rc.callback != null)
 				rc.callback.executed(rrc.clientRequest, false);
 			else if (rrc.callback != null)
@@ -446,7 +474,7 @@ public class PaxosManager<NodeIDType> {
 											- PaxosManager.this.outstanding.lastIncremented > REQUEST_TIMEOUT)
 									|| monitorIterval > 0) {
 								HashMap<Long, String> instances = new HashMap<Long, String>();
-								for (PaxosConfig.RequestAndCallback rc : PaxosManager.this.outstanding.requests
+								for (RequestAndCallback rc : PaxosManager.this.outstanding.requests
 										.values())
 									instances.put(rc.requestPacket.requestID,
 											rc.requestPacket.getPaxosID() + ":"
@@ -467,7 +495,7 @@ public class PaxosManager<NodeIDType> {
 								if (!PaxosManager.this.outstanding.requests
 										.isEmpty()
 										&& PaxosManager.this.outstanding.requests instanceof GCConcurrentHashMap)
-									((GCConcurrentHashMap<Long, PaxosConfig.RequestAndCallback>) PaxosManager.this.outstanding.requests)
+									((GCConcurrentHashMap<Long, RequestAndCallback>) PaxosManager.this.outstanding.requests)
 											.tryGC(REQUEST_TIMEOUT);
 
 							}
@@ -1178,7 +1206,7 @@ public class PaxosManager<NodeIDType> {
 					new Object[] { this, pism.getPaxosIDVersion(),
 							requestPacket.getSummary() });
 
-			this.outstanding.enqueue(new PaxosConfig.RequestAndCallback(requestPacket,
+			this.outstanding.enqueue(new RequestAndCallback(requestPacket,
 					callback));
 			this.handleIncomingPacket(requestPacket);
 		} else
@@ -1548,11 +1576,11 @@ public class PaxosManager<NodeIDType> {
 		request.setEntryReplicaAndReturnCount(this.myID);
 
 		// if (request.getEntryReplica() == getMyID())
-		this.outstanding.enqueue(new PaxosConfig.RequestAndCallback(request, null));
+		this.outstanding.enqueue(new RequestAndCallback(request, null));
 		if (request.batchSize() > 0)
 			for (RequestPacket req : request.getBatched())
 				// if (request.getEntryReplica() == getMyID())
-				this.outstanding.enqueue(new PaxosConfig.RequestAndCallback(req, null));
+				this.outstanding.enqueue(new RequestAndCallback(req, null));
 		if (Util.oneIn(10))
 			DelayProfiler.updateMovAvg("outstanding",
 					this.outstanding.requests.size());
@@ -3334,7 +3362,7 @@ public class PaxosManager<NodeIDType> {
 	}
 
 	protected MessageDigest getMessageDigest() {
-		return PendingDigests.getMessageDigest();
+		return RequestPacket.getMessageDigest();
 	}
 
 	private static final int DIGEST_THRESHOLD = Config
@@ -3441,7 +3469,7 @@ public class PaxosManager<NodeIDType> {
 	}
 
 	protected AcceptPacket getPreviouslyIssuedAccept(RequestPacket proposal) {
-		PaxosConfig.RequestAndCallback rc = null;
+		RequestAndCallback rc = null;
 		if (((rc = this.outstanding.requests.get(proposal.getRequestID())) != null || (rc = this.outstanding.conflictIDRequests
 				.get(proposal)) != null)
 				&& rc.requestPacket instanceof AcceptPacket
@@ -3452,7 +3480,7 @@ public class PaxosManager<NodeIDType> {
 
 	protected AcceptPacket setIssuedAccept(AcceptPacket accept) {
 		assert(accept==null || accept.getRequestValue()!=null) : accept;
-		PaxosConfig.RequestAndCallback rc = accept!=null ? this.outstanding.requests.get(accept
+		RequestAndCallback rc = accept!=null ? this.outstanding.requests.get(accept
 				.getRequestID()) : null;
 		if (rc != null
 				|| (rc = this.outstanding.conflictIDRequests.get(accept)) != null)
