+++ src/edu/umass/cs/reconfiguration/reconfigurationpackets/BatchedCreateServiceName.java
@@ -1,19 +1,11 @@
 package edu.umass.cs.reconfiguration.reconfigurationpackets;
 
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-
+import edu.umass.cs.nio.interfaces.Stringifiable;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import edu.umass.cs.nio.interfaces.Stringifiable;
-import edu.umass.cs.reconfiguration.reconfigurationutils.ConsistentReconfigurableNodeConfig;
-import edu.umass.cs.utils.Util;
+import java.net.InetSocketAddress;
 
 /**
  * @author arun
@@ -34,8 +26,9 @@ public class BatchedCreateServiceName extends CreateServiceName {
 	 * @param name
 	 * @param epochNumber
 	 * @param state
+	 * @TODO Change access to private after moving test to test source
 	 */
-	private BatchedCreateServiceName(InetSocketAddress initiator, String name,
+	public BatchedCreateServiceName(InetSocketAddress initiator, String name,
 			int epochNumber, String state) {
 		super(initiator, name, epochNumber, state);
 		this.initialStates = new String[1];
@@ -50,8 +43,9 @@ public class BatchedCreateServiceName extends CreateServiceName {
 	 * @param names
 	 * @param epochNumber
 	 * @param states
+	 * @TODO Change access to private after moving test to test source
 	 */
-	private BatchedCreateServiceName(InetSocketAddress initiator,
+	public BatchedCreateServiceName(InetSocketAddress initiator,
 			String[] names, int epochNumber, String[] states) {
 		super(initiator, names[0], epochNumber, states[0]);
 		assert (epochNumber == 0);
@@ -109,44 +103,4 @@ public class BatchedCreateServiceName extends CreateServiceName {
 		return super.getSummary() + ":|batched|=" + this.names.length;
 	}
 
-	public static void main(String[] args) {
-		try {
-			Util.assertAssertionsEnabled();
-			InetSocketAddress isa = new InetSocketAddress(
-					InetAddress.getByName("localhost"), 2345);
-			int numNames = 1000;
-			String[] reconfigurators = { "RC43", "RC22", "RC78", "RC21",
-					"RC143" };
-			String namePrefix = "someName";
-			String defaultState = "default_initial_state";
-			String[] names = new String[numNames];
-			String[] states = new String[numNames];
-			for (int i = 0; i < numNames; i++) {
-				names[i] = namePrefix + i;
-				states[i] = defaultState + i;
-			}
-			BatchedCreateServiceName bcreate1 = new BatchedCreateServiceName(
-					isa, "random0", 0, "hello");
-			BatchedCreateServiceName bcreate2 = new BatchedCreateServiceName(
-					isa, names, 0, states);
-			System.out.println(bcreate1.toString());
-			System.out.println(bcreate2.toString());
-
-			// translate a batch into consistent constituent batches
-			Collection<Set<String>> batches = ConsistentReconfigurableNodeConfig
-					.splitIntoRCGroups(
-							new HashSet<String>(Arrays.asList(names)),
-							new HashSet<String>(Arrays.asList(reconfigurators)));
-			int totalSize = 0;
-			int numBatches = 0;
-			for (Set<String> batch : batches)
-				System.out.println("batch#" + numBatches++ + " of size "
-						+ batch.size() + " (totalSize = "
-						+ (totalSize += batch.size()) + ")" + " = " + batch);
-			assert(totalSize==numNames);
-			System.out.println(bcreate2.getSummary());
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
 }
